+++ wxt.config.ts
@@ -7,7 +7,12 @@ import { defineConfig } from 'wxt';
 // See https://wxt.dev/api/config.html
 export default defineConfig({
   extensionApi: 'chrome',
-  modules: ['@wxt-dev/module-react'],
+  modules: ['@wxt-dev/module-react','@wxt-dev/unocss'],
+  unocss:{
+    configOrPath:{
+      mode: 'shadow-dom',
+    }
+  },
   manifest: {
     action: {
       default_title: "Click to open admin page"
@@ -85,5 +90,4 @@ export default defineConfig({
       }
     }
   }
-
 });